+++ src/app/app.component.ts
@@ -1,4 +1,5 @@
 import {Component} from '@angular/core';
+import {NgbDatepickerConfig} from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'lfb-root',
@@ -7,4 +8,9 @@ import {Component} from '@angular/core';
 })
 export class AppComponent {
   title = 'formbuilder-lhcforms';
+  constructor(private dateConfig: NgbDatepickerConfig) {
+    const today = new Date();
+    dateConfig.minDate = {year: today.getUTCFullYear() - 100, month: today.getUTCMonth()+1, day: today.getUTCDate()};
+    dateConfig.maxDate = {year: today.getUTCFullYear() + 100, month: today.getUTCMonth()+1, day: today.getUTCDate()};
+  }
 }