+++ webpack.config.js
@@ -1,0 +1,20 @@
+const path = require('path');
+require('json5/lib/register');
+
+module.exports = {
+  // ...
+  module: {
+    rules: [
+      // ...
+      {
+        test: /\.json5$/,
+        use: [
+          {
+            loader: path.resolve(__dirname, 'json5-loader.js'),
+          },
+        ],
+      },
+    ],
+  },
+  // ...
+};